🐛 handle bad data in 2023 8 solve_1
The problem was to decide how to handle this case.
The CLI already makes it possible to select what example to use,
so adding complexity to use different example files by default is unnecessary.
The real issue is when the first solution of a problem is not compatible with more complex data.
It seems to be the better approach to return a readable message in this case, to not prevent execution of part 2.